integrating git feature flags
